# Fix "common_part" selection in server