--- a/examples/server/utils.hpp
+++ b/examples/server/utils.hpp
@@ -339,8 +339,6 @@ static size_t common_part(const llama_context * ctx, const llama_model * model, 
     return token_idx;
 }
 
-
-
 static bool ends_with(const std::string & str, const std::string & suffix) {
     return str.size() >= suffix.size() && 0 == str.compare(str.size() - suffix.size(), suffix.size(), suffix);
 }