This is my attempt to address https://github.com/ggml-org/llama.cpp/issues/11970 

It has some similarities to https://github.com/ggml-org/llama.cpp/pull/12067 but is not a port it is implemented differently.

It matches tokens in cache directly to the string passed in the prompt.

I do not know how to reliably reproduce the issue so I'm not sure if it works.

@vnicolici Do you mind testing?