@saood06 2025-09-03
>But I wonder if it wouldn't be better to first try the tokens directly.  If the full prompt is matched, there would be no need to do string matching, and the behavior will be the same as before. 

I can see how that might result in better performance for the average user.

>If we find a shorter match, one can do string matching from the point of disagreement.

But I don't think starting from the point of disagreement makes sense because if you compare tokens directly and everything matches you don't need to detokenize anything. In order to know where the point of disagreement is on the string wouldn't you need to do string matching.