@ikawrakow 2025-09-03
So, my understanding is that the `tokens -> text` conversion is not necessarily reversible, i.e., after `text -> tokens'`, `tokens'` may not be the same as `tokens`. But if `tokens[j] == tokens'[j]` up to `j_match`, then one can convert `tokens'[j_match...end]` to string and do string matching from there. No? This may not be more efficient than just matching strings as we now need two conversions (`text -> tokens'` and then `tokens'[j_match...] -> text'`), but is guaranteed to be not worse than the existing implementation.

Alternatively one can match tokens and strings, and one takes the result with the greater matched length.

All of this is only to avoid a potential failure mode in the string matching approach that we both don't see.