@usrlocalben 2025-10-05
I tried this out. I finally hit a prompt that could benefit from this because the cache & tokenizer disagree at 20KT deep into a 100KT prompt.

I noticed that it refuses to reuse the prompt, starting at p0=0 each time.

I added some debug logging in the new common_part() and was surprised to see that the very first token (which I expected to be BOS or some special chat template token) was a token from well past the beginning of the prompt.

Could it be related to fragmentation? 🤔