@usrlocalben 2025-10-05
Thinking a moment more, I realize this may be unrelated to this patch as my sample doesn't have any output tokens, just 100K input tokens, so it shouldn't be due to the tokenizer/decode mismatch problem.