@usrlocalben 2025-10-05
I compared with main/HEAD as of now, and it appears to be a fault/regression in this PR.

my figures were off wrt. tokens but the gist is the same.

these are cold-starts / re-gens pairs for the same prompt

main/HEAD
```
INFO [            update_slots] kv cache rm [p0, end) | tid="140196405805056" timestamp=1759694453 id_slot=0 id_task=2 p0=65536
INFO [            update_slots] kv cache rm [p0, end) | tid="140196405805056" timestamp=1759694490 id_slot=0 id_task=2 p0=69632
INFO [            update_slots] kv cache rm [p0, end) | tid="140196405805056" timestamp=1759694528 id_slot=0 id_task=2 p0=73728
INFO [           print_timings] prompt eval time     =  580853.17 ms / 76425 tokens (    7.60 ms per token,   131.57 tokens per second) | tid="140196405805056" timestamp=1759694583 id_slot=0 id_task=2 t_prompt_processing=580853.171 n_prompt_tokens_processed=76425 t_token=7.600303186130192 n_tokens_second=131.5736985104623
INFO [           print_timings] generation eval time =   30540.66 ms /   403 runs   (   75.78 ms per token,    13.20 tokens per second) | tid="140196405805056" timestamp=1759694583 id_slot=0 id_task=2 t_token_generation=30540.659 n_decoded=403 t_token=75.7832729528536 n_tokens_second=13.195524038954103
INFO [           print_timings]           total time =  611393.83 ms | tid="140196405805056" timestamp=1759694583 id_slot=0 id_task=2 t_prompt_processing=580853.171 t_token_generation=30540.659 t_total=611393.83
INFO [            update_slots] slot released | tid="140196405805056" timestamp=1759694583 id_slot=0 id_task=2 n_ctx=131072 n_past=76827 n_system_tokens=0 n_cache_tokens=76827 truncated=true
INFO [            update_slots] all slots are idle | tid="140196405805056" timestamp=1759694583
INFO [      log_server_request] request | tid="140186308542464" timestamp=1759694583 remote_addr="127.0.0.1" remote_port=44778 status=200 method="POST" path="/v1/chat/completions" params={}
INFO [            update_slots] all slots are idle | tid="140196405805056" timestamp=1759694583
Grammar: 
Grammar lazy: false
Chat format: Content-only
INFO [   launch_slot_with_task] slot is processing task | tid="140196405805056" timestamp=1759694613 id_slot=0 id_task=425
INFO [            update_slots] we have to evaluate at least 1 token to generate logits | tid="140196405805056" timestamp=1759694613 id_slot=0 id_task=425
INFO [            update_slots] kv cache rm [p0, end) | tid="140196405805056" timestamp=1759694613 id_slot=0 id_task=425 p0=76424
INFO [           print_timings] prompt eval time     =     259.95 ms /     1 tokens (  259.95 ms per token,     3.85 tokens per second) | tid="140196405805056" timestamp=1759694639 id_slot=0 id_task=425 t_prompt_processing=259.954 n_prompt_tokens_processed=1 t_token=259.954 n_tokens_second=3.8468344399393737
INFO [           print_timings] generation eval time =   25420.14 ms /   336 runs   (   75.66 ms per token,    13.22 tokens per second) | tid="140196405805056" timestamp=1759694639 id_slot=0 id_task=425 t_token_generation=25420.139 n_decoded=336 t_token=75.65517559523809 n_tokens_second=13.217866353917263
INFO [           print_timings]           total time =   25680.09 ms | tid="140196405805056" timestamp=1759694639 id_slot=0 id_task=425
```


pr
```
INFO [            update_slots] kv cache rm [p0, end) | tid="140172995358720" timestamp=1759692609 id_slot=0 id_task=2 p0=57344
INFO [            update_slots] kv cache rm [p0, end) | tid="140172995358720" timestamp=1759692643 id_slot=0 id_task=2 p0=61440
INFO [            update_slots] kv cache rm [p0, end) | tid="140172995358720" timestamp=1759692677 id_slot=0 id_task=2 p0=65536
INFO [            update_slots] kv cache rm [p0, end) | tid="140172995358720" timestamp=1759692714 id_slot=0 id_task=2 p0=69632
INFO [            update_slots] kv cache rm [p0, end) | tid="140172995358720" timestamp=1759692752 id_slot=0 id_task=2 p0=73728
INFO [           print_timings] prompt eval time     =  579941.19 ms / 76425 tokens (    7.59 ms per token,   131.78 tokens per second) | tid="140172995358720" timestamp=1759692799 id_slot=0 id_task=2 t_prompt_processing=579941.191 n_prompt_tokens_processed=76425 t_token=7.588370179914949 n_tokens_second=131.7806032508562
INFO [           print_timings] generation eval time =   23004.48 ms /   303 runs   (   75.92 ms per token,    13.17 tokens per second) | tid="140172995358720" timestamp=1759692799 id_slot=0 id_task=2 t_token_generation=23004.476 n_decoded=303 t_token=75.92236303630362 n_tokens_second=13.171349784276766
INFO [           print_timings]           total time =  602945.67 ms | tid="140172995358720" timestamp=1759692799 id_slot=0 id_task=2 t_prompt_processing=579941.191 t_token_generation=23004.476 t_total=602945.667
INFO [      log_server_request] request | tid="140162879156224" timestamp=1759692799 remote_addr="127.0.0.1" remote_port=34622 status=200 method="POST" path="/v1/chat/completions" params={}
INFO [            update_slots] slot released | tid="140172995358720" timestamp=1759692799 id_slot=0 id_task=2 n_ctx=131072 n_past=76727 n_system_tokens=0 n_cache_tokens=76727 truncated=true
INFO [            update_slots] all slots are idle | tid="140172995358720" timestamp=1759692799
Grammar: 
Grammar lazy: false
Chat format: Content-only
INFO [   launch_slot_with_task] slot is processing task | tid="140172995358720" timestamp=1759692831 id_slot=0 id_task=325
INFO [            update_slots] kv cache rm [p0, end) | tid="140172995358720" timestamp=1759692831 id_slot=0 id_task=325 p0=0
INFO [            update_slots] kv cache rm [p0, end) | tid="140172995358720" timestamp=1759692857 id_slot=0 id_task=325 p0=4096
INFO [            update_slots] kv cache rm [p0, end) | tid="140172995358720" timestamp=1759692884 id_slot=0 id_task=325 p0=8192
INFO [            update_slots] kv cache rm [p0, end) | tid="140172995358720" timestamp=1759692911 id_slot=0 id_task=325 p0=12288
INFO [            update_slots] kv cache rm [p0, end) | tid="140172995358720" timestamp=1759692938 id_slot=0 id_task=325 p0=16384
INFO [            update_slots] kv cache rm [p0, end) | tid="140172995358720" timestamp=1759692966 id_slot=0 id_task=325 p0=20480
```